// core/src/types/supabase.ts
export interface definitions {
  signal: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    value: number;
    type: string;
    user: string;
  };
  alert: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    /**
     * Note:
     * This is a Foreign Key to `insight.id`.<fk table='insight' column='id'/>
     */
    insight: number;
    acknowledge?: boolean;
    user: string;
    severity: number;
  };
  insight: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    value?: { [key: string]: any };
    user: string;
  };
}